requirements.txt
flask
requests
python-dateutil
gunicorn
tzdata